Add check for a ZERO value (#104)
The JS language considers the value ZERO to be
"false". This has to do with the fact that the
numeric representation of true/false are 1/0.

This was of course preventing the initial value
of a numeric field to be set to zero since there
was a check on the mounted function that saw if
the value was defined.

I've added a simple OR case, where if the value
was evaluated to FALSE than we check again for
if the value is zero. This fixed the case where
the value ZERO was not being shown when it was
deliberately chosen.

It did break another case when we pass an empty
string. To fix it I also test if the original value
is not empty (which would be evaluated to zero)

Co-authored-by: Flavio Diez <f.diez@right-basedonscience.de>